# Back/routes/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"TypeConge": {
				"enum": [
					"ANNUEL",
					"MALADIE",
					"MATERNITE",
					"PATERNITE",
					"EXCEPTIONNEL",
					"AUTRE"
				],
				"type": "string"
			},
			"CongeOutput": {
				"properties": {
					"id_conge": {
						"type": "number",
						"format": "double"
					},
					"matricule": {
						"type": "string"
					},
					"motif": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/TypeConge"
					},
					"nbr_jours_permis": {
						"type": "number",
						"format": "double"
					},
					"solde_conge": {
						"type": "number",
						"format": "double"
					},
					"date_depart": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"date_reprise": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"personne_interim": {
						"type": "string"
					},
					"user": {
						"properties": {
							"email": {
								"type": "string"
							},
							"prenom": {
								"type": "string"
							},
							"nom": {
								"type": "string"
							}
						},
						"required": [
							"email",
							"prenom",
							"nom"
						],
						"type": "object"
					},
					"userInterim": {
						"properties": {
							"matricule": {
								"type": "string"
							},
							"prenom": {
								"type": "string"
							},
							"nom": {
								"type": "string"
							}
						},
						"required": [
							"matricule",
							"prenom",
							"nom"
						],
						"type": "object"
					}
				},
				"required": [
					"id_conge",
					"matricule",
					"motif",
					"type",
					"nbr_jours_permis",
					"solde_conge",
					"date_depart",
					"date_reprise"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CongeValidationResult": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CongeCreateParams": {
				"properties": {
					"matricule": {
						"type": "string"
					},
					"motif": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/TypeConge"
					},
					"nbr_jours_permis": {
						"type": "number",
						"format": "double"
					},
					"solde_conge": {
						"type": "number",
						"format": "double"
					},
					"date_depart": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"date_reprise": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"personne_interim": {
						"type": "string"
					}
				},
				"required": [
					"matricule",
					"motif",
					"type",
					"nbr_jours_permis",
					"solde_conge",
					"date_depart",
					"date_reprise"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CongeUpdateParams": {
				"properties": {
					"matricule": {
						"type": "string"
					},
					"motif": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/TypeConge"
					},
					"nbr_jours_permis": {
						"type": "number",
						"format": "double"
					},
					"solde_conge": {
						"type": "number",
						"format": "double"
					},
					"date_depart": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"format": "date-time"
							}
						]
					},
					"date_reprise": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"format": "date-time"
							}
						]
					},
					"personne_interim": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"SoldeValidationResult": {
				"properties": {
					"valid": {
						"type": "boolean"
					},
					"soldeActuel": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"valid",
					"soldeActuel",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "back-rh-connect",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/conges/{id}": {
			"get": {
				"operationId": "GetConge",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CongeOutput"
								}
							}
						}
					}
				},
				"tags": [
					"Conge"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateConge",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CongeValidationResult"
								}
							}
						}
					}
				},
				"tags": [
					"Conge"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH",
							"superviseur"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CongeUpdateParams"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteConge",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CongeValidationResult"
								}
							}
						}
					}
				},
				"tags": [
					"Conge"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/conges": {
			"post": {
				"operationId": "CreateConge",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CongeValidationResult"
								}
							}
						}
					}
				},
				"tags": [
					"Conge"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CongeCreateParams"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllConges",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CongeOutput"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Conge"
				],
				"security": [],
				"parameters": []
			}
		},
		"/conges/matricule/{matricule}": {
			"get": {
				"operationId": "GetCongesByMatricule",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CongeOutput"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Conge"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "matricule",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/conges/type/{type}": {
			"get": {
				"operationId": "GetCongesByType",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CongeOutput"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Conge"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH",
							"superviseur"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "type",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/TypeConge"
						}
					}
				]
			}
		},
		"/conges/motif/{motif}": {
			"get": {
				"operationId": "GetCongesByMotif",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CongeOutput"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Conge"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH",
							"superviseur"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "motif",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/conges/en-cours": {
			"get": {
				"operationId": "GetCongesEnCours",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CongeOutput"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Conge"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH",
							"superviseur"
						]
					}
				],
				"parameters": []
			}
		},
		"/conges/a-venir": {
			"get": {
				"operationId": "GetCongesAVenir",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CongeOutput"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Conge"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH",
							"superviseur"
						]
					}
				],
				"parameters": []
			}
		},
		"/conges/periode": {
			"get": {
				"operationId": "GetCongesByPeriode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CongeOutput"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Conge"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH",
							"superviseur"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "dateDebut",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "dateFin",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/conges/solde/{matricule}": {
			"get": {
				"operationId": "GetSoldeConge",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"solde": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"solde"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Conge"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "matricule",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/conges/valider-solde": {
			"post": {
				"operationId": "ValiderSoldeConge",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SoldeValidationResult"
								}
							}
						}
					}
				},
				"tags": [
					"Conge"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH",
							"superviseur"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"joursDemanDes": {
										"type": "number",
										"format": "double"
									},
									"matricule": {
										"type": "string"
									}
								},
								"required": [
									"joursDemanDes",
									"matricule"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/conges/statistiques/{matricule}": {
			"get": {
				"operationId": "GetStatistiquesConge",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"soldeActuel": {
											"type": "number",
											"format": "double"
										},
										"totalJours": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"soldeActuel",
										"totalJours"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Conge"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH",
							"superviseur"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "matricule",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "annee",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}